Added wrapper to API calls to fetch all pages with the data
Now fetching all pages, also filtering to get the data for just Balkan
countries. Async is used but currently provides no benefit because we
cannot know for larger sets how much data is there, but have to wait for
an empty response. Plan is to fetch in batches so that the data can be
fetched faster.