Implemented RateLimiter class
RateLimiter class will be used to make sure we do not go over the
request limit that openAQ API has. The solution will work in all
scenarios with concurrent execution, only minor issue is that when
requesting a batch fetch the batch will reserve some requests without
using them which could lead to a small performance loss. So for example
if we have used 57/60 requests, and a batch wants to make 5 more it will
block all other requests, even if more efficient would be to for
example let 3 single request task fire and then take its turn. If we
just prioritize smaller requests they could starve the batches, so the
only solution i can think of is to create a queue of some sorts maybe.